requirements.txt
backports.tarfile==1.2.0
chill==0.10.1
colorama==0.4.6
importlib-metadata==8.0.0
importlib-resources==6.4.0
ipykernel==6.29.5
jaraco.collections==5.1.0
matplotlib==3.9.2
networkx==3.4.2
nltk==3.9.1
openai==1.54.4
pandas==2.2.3
pip-chill==1.0.3
pipreqs==0.5.0
python-dotenv==1.0.1
scipy==1.14.1
tomli==2.0.1
transformers==4.46.2
tree-sitter==0.23.2
tree-sitter-python==0.23.4
